move the completion parsing logic to `ollama.ts`
